add indeterminism to FindNearestFreePosition and clean float conversions

+++ pkg/utils/math.go
@@ -18,8 +18,8 @@ func CalculateMovementCost(dx, dy float64) float64 {
 	return 1.0 // Horizontal/vertical movement
 }
 
-func EuclideanDistance(c1, c2 [2]int) float64 {
-	dx := float64((c1[0] - c2[0]))
-	dy := float64((c1[1] - c2[1]))
+func EuclideanDistance(c1, c2 [2]float64) float64 {
+	dx := (c1[0] - c2[0])
+	dy := (c1[1] - c2[1])
 	return math.Sqrt(dx*dx + dy*dy)
 }
